Make back button in 2nd activity close app
I made first activity to finish itself after calling the second.
So now the default behaviour of back button in second activity
closes the app window.

Of course the app still stays in the list of active apps,
I see it after pressing the square button. But I think that
my way to «close» the app after the back button press is
good enought because conforms with google ideology of not
killing apps.